requirements.txt
git+https://github.com/openai/CLIP
git+https://github.com/crowsonkb/guided-diffusion
git+https://github.com/CompVis/taming-transformers
ftfy
regex
tqdm
omegaconf
pytorch-lightning
kornia
einops
transformers
stegano
python-xmp-toolkit
imgtag
pillow
taming-transformers
imageio-ffmpeg
torch==1.10.0
torchvision==0.11.1
torchaudio===0.10.0
realesrgan
moviepy